Move the exporters to the I/O module
