--- battdat/exporters/__init__.py
@@ -1,1 +1,0 @@
-"""Tools for exporting battery data into a format usable by other tools"""
